% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_1tot1_orthologs.r
\name{check_1to1_orthologs}
\alias{check_1to1_orthologs}
\title{check.1to1.orthologs}
\usage{
check_1to1_orthologs(orth)
}
\arguments{
\item{orthologs}{Data table with MGI.symbol and HGNC.symbol columns. Should only contain 1:1 homologs}
}
\description{
\code{check.1to1.orthologs} Confirms that the datatable passed as input has columns for MGI.symbol, HGNC.symbol and only contains 1:1 orthologs
}
\examples{
check.1to1.orthologs(orth)
}
