% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_datasets.r
\name{compare_datasets}
\alias{compare_datasets}
\title{compare.datasets}
\usage{
compare_datasets(dataset1, dataset2, reps = 1000, geneListHGNC = NULL,
  geneListMGI = NULL, orthologs, sharedName, pSides = "onesided")
}
\arguments{
\item{dataset1}{Cell type data, loaded by prepare.specificity.comparison.across.datasets()}

\item{dataset2}{Cell type data file (formatted as expected for EWCE CTD files)}

\item{reps}{Number of replacates to sample}

\item{geneListHGNC}{Gene list (HGNC symbols) to test the specificity of between the two datasets}

\item{orthologs}{Data table with MGI.symbol and HGNC.symbol columns. Should only contain 1:1 homologs}

\item{sharedName}{A celltype name that is in both ctd specificity files}

\item{pSides}{Either 'onesided' or 'twosided' depending on whether you only care about enrichment in celltypes, or depletion as well}

\item{dataset1$datasetName}{The 'title' of the dataset to be used in figures}

\item{col2}{The name of the celltype to use from dataset2}

\item{dataset2$datasetName}{The 'title' of the dataset to be used in figures}

\item{level2annot}{Array of cell types, with each sequentially corresponding a column in the expression matrix}
}
\value{
p Probility (based on bootstrapping) that there is a difference in the specificity of the gene list between the two datasets

z Z-score for the difference in the specificity of the gene list between the two datasets
}
\description{
\code{compare.datasets} Takes two cell type data files and tests whether the genes contained in 'geneListHGNC' are more specific to one dataset
than the other. This is restricted to the cell types named in sbaredName
}
\examples{
a=1+1
}
