% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_publication_plot_for_single_list.r
\name{generate_publication_plot_for_single_list}
\alias{generate_publication_plot_for_single_list}
\title{generate_publication_plot_for_single_list}
\usage{
generate_publication_plot_for_single_list(allRes,
  listName = "dendrite_enriched_transcripts_HGNC_1to1only")
}
\arguments{
\item{allRes}{Output of test_all_comparisons}

\item{listName}{Used to select a particular list from listName (not relevant to current version of package)}
}
\value{
thePlot (as a ggplot)
}
\description{
\code{generate_publication_plot_for_single_list} Generates the main plot
}
\examples{
allDataSets = list(mouseSS_dist,tasic_mouse_dist,dronc_mouse_dist,dronc_human_dist,divseq_dist)
res = test_all_comparisons(allDataSets,orthologs,sharedName="Pyramidal Neuron")
thePlot = generate_publication_plot_for_single_list(res)
}
