% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_comparisons.r
\name{get_group_comparisons}
\alias{get_group_comparisons}
\title{get_group_comparisons}
\usage{
get_group_comparisons(datasets)
}
\arguments{
\item{datasets}{Set of datasets}
}
\value{
comparisons
}
\description{
\code{get_group_comparisons} Internal function
}
\examples{
# 1==1
}
