% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genelist.r
\name{load_genelist}
\alias{load_genelist}
\title{load_genelist}
\usage{
load_genelist(list_path, lookuptable, speciesWanted = "human")
}
\arguments{
\item{list_path}{Folder where it is stored}

\item{lookuptable}{Dataframe containing HGNC.symbol and MGI.symbol as columns}

\item{speciesWanted}{Either 'human' or 'mouse'}
}
\value{
res Results
}
\description{
\code{load_genelist} Loads a gene list from a file
}
\details{
Expects the first line to be descriptive
}
\examples{
# list_path = sprintf("\%s/\%s.txt",path,listN)
# geneListHGNC = load.genelist(list_path,orthologs,speciesWanted="human")
}
