% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_all_comparisons.r
\name{test_all_comparisons}
\alias{test_all_comparisons}
\title{test_all_comparisons}
\usage{
test_all_comparisons(allDataSets, orthologs,
  sharedName = "Pyramidal Neuron")
}
\arguments{
\item{allDataSets}{Created using prepare_specificity_comparison_across_datasets}

\item{orthologs}{Dataframe containing HGNC.symbol and MGI.symbol as columns}

\item{sharedName}{The cell type being compared across datasets}
}
\value{
res Results
}
\description{
\code{test_all_comparisons} Runs the main dendritic depletion analysis
}
\examples{
allDataSets = list(mouseSS_dist,tasic_mouse_dist,dronc_mouse_dist,dronc_human_dist,divseq_dist)
res = test_all_comparisons(allDataSets,orthologs,sharedName="Pyramidal Neuron")
}
